Move filter files to single exports
