+++ docs/prd/16-filter-rework/notes.md
@@ -1,12 +1,17 @@
-# Filter rework notes
+## Filter rework notes
 
-- Update DataLayerProvider to use the new filter functions
-- Data table needs to be case insensitive
-- Data table should include option to globally filter vs. just local table
-- Data table global search seems to not work
-- Pivot table - not clear how to activate filters
+### Pivot Table
 
-## Questions
+- Not clear how to activate filters - click the search icon - should change to a filter icon
+- Should highlight the active filters in the pivot table
+- No obvious way to engage a row filter
+- Remove the `toast` that appears when a filter is applied
 
-- Curious how the create + update + apply functions will work
-- DataTable got wrecked by filter updates - will need to revert and do manually
+### Related to Data Table
+
+- Needs to be case insensitive for contains and other text filters
+- Should include option to filter only the component or to filter the entire crossfilter context
+- Global search seems to not work - does it search all fields? or just those involved in the display
+- Add highlighter when using the global search or other filters
+- Show filter as a row below the header - add option to change type
+- Add a popover for quick access to advanced filtering options
