+++ src/components/charts/ScatterPlot/definition.ts
@@ -5,7 +5,8 @@ import { ScatterChart } from "lucide-react";
 import { ScatterPlotSettingsPanel } from "./ScatterPlotSettingsPanel";
 import { ScatterPlot } from "./ScatterPlot";
 import { IdType } from "@/providers/DataLayerProvider";
-import { applyFilter, getAxisFilter } from "@/hooks/useFilters";
+import { getAxisFilter } from "@/hooks/getAxisFilter";
+import { applyFilter } from "@/hooks/applyFilter";
 import { Filter } from "@/types/FilterTypes";
 
 export interface ScatterPlotSettings extends BaseChartSettings {
